Refactors auction data storage and processing (#74)
## Type of change
- [x] ✨ Feature
- [ ] 🐛 Bug fix
- [x] 🧹 Code refactor
- [x] 🧪 Tests
- [ ] 📚 Documentation
- [x] 🔧 Build/CI
- [ ] 🔄 Other (describe below)

## What has been done?
- **Database Schema Redesign:** Replaced legacy auction-related tables
(`auction_item_modifier_link`, `auction_item_modifier`, `auction_item`,
and `auction`) with a new, optimized schema. Introduces two new tables:
- `auction`: Stores unique item variants and aggregated statistics,
including `p25` and `p75` percentile buyouts, and uses a combined ID
based on the connected realm and item variant.
- `auction_price`: Stores individual auction listings, linked to the
`auction` table, preserving historical price data.
- **Optimized Auction Persistence Pipeline:**
- Overhauled `AuctionSnapshotPersistenceService` to efficiently stream
and group incoming auction data by unique item variants, calculate `p25`
and `p75` percentiles, and then persist the new `Auction` and
`AuctionPrice` records.
- Refactored `AuctionJDBCRepository` to leverage multi-row `INSERT ...
ON DUPLICATE KEY UPDATE` statements for faster bulk upserts of the new
domain objects.
- **Streamlined Hourly Statistics Generation:**
- Simplified `AuctionStatsHourlyService` logic, now delegating to
`AuctionStatsHourlyJDBCRepository` which uses direct `INSERT ... SELECT`
queries from the new `auction` table to aggregate hourly statistics.
This eliminates in-memory grouping and reduces processing overhead.
- **Enhanced Market Search Capabilities:**
- Updated `AuctionMarketSearchRepository` and
`CraftingMarketSearchRepository` to query the new `auction` table
directly for current price data.
- Integrated `p25Price` and `p75Price` fields into market search results
and DTOs, providing more robust pricing analytics.
- **Frontend UI Updates:**
- Modified frontend UI models, mappers, and components to display the
newly available `p25` and `p75` percentile prices in both general market
and crafting search results.
- **Removed Legacy Code:** Deleted unused repositories and utility
classes that are no longer compatible with the new data model, reducing
code complexity.
- **Build Fix:** Corrected a `pom.xml` configuration to remove an
incorrect source directory.

## Motivation / Context
This major refactor aims to significantly improve the efficiency and
scalability of auction data processing and storage. By separating
auction item variants from individual listings and optimizing batch
inserts, the system can handle large volumes of auction data more
effectively. The introduction of `p25` and `p75` percentile prices
directly in the `auction` table and the market search results provides
users with more robust pricing analytics, enhancing decision-making for
buyers and sellers. The move to direct database queries for hourly
statistics further reduces application memory footprint and processing
time, aligning with the goal to optimize the insertion of all current
prices.

## Screenshots (if applicable)


## Checklist
- [x] I’ve tested this manually
- [x] I’ve added tests if applicable
- [ ] I’ve updated docs if needed
- [x] I’ve followed the coding conventions

## Related Issues
Relates to #55

+++ backend/src/main/kotlin/net/jonasmf/auctionengine/dto/auction/AuctionDTO.kt
@@ -7,7 +7,7 @@ import net.jonasmf.auctionengine.constant.AuctionTimeLeft
 data class AuctionDTO(
     val id: Long,
     val item: AuctionItemDTO,
-    val quantity: Long,
+    val quantity: Int,
     val bid: Long? = null,
     val unit_price: Long?, // Commodity price
     val buyout: Long?, // Realm price
